% Generated by roxygen2 (4.0.2): do not edit by hand
\name{angleRotation}
\alias{angleRotation}
\title{Rotate a letter according to a set of angle.}
\usage{
angleRotation(letter, angle)
}
\arguments{
\item{letter}{}

\item{angle}{}
}
\description{
\code{angleRotation} rotates a letter according to it's width, i.e.,
it minimizes or maximizes it (for the letters M & W).
}

