% Generated by roxygen2 (4.0.2): do not edit by hand
\name{countChar}
\alias{countChar}
\title{Count number of characters}
\usage{
countChar(input_string, char_universe = letters)
}
\arguments{
\item{input_string}{input string}

\item{char_universe}{Possible set of characters}
}
\description{
\code{countChar}
}

