% Generated by roxygen2 (4.0.2): do not edit by hand
\name{img2Black}
\alias{img2Black}
\title{Transforms an image to black & white}
\usage{
img2Black(img)
}
\arguments{
\item{img}{A matrix of imagedata resulting from a readJpeg import.}
}
\description{
\code{img2Black} transforms a multilayered and multicolored
captcha into a black & white image. The values are (rgb = 0) for black and
(rbg = 255) for white.
}
\examples{
captcha <- img2Black(captcha)
}

