% Generated by roxygen2 (4.0.2): do not edit by hand
\name{img2Grey}
\alias{img2Grey}
\title{Transforms an image to Greyscale}
\usage{
img2Grey(img)
}
\arguments{
\item{img}{An image resulting from a readImage import.}
}
\description{
\code{img2Grey} transforms a multilayered and multicolored
Image into a Greyscale image.
}
\examples{
captcha <- img2Grey(captcha)
}

