% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isWhite}
\alias{isWhite}
\title{Find parts of an image that is white}
\usage{
isWhite(img_data, axes = "v")
}
\arguments{
\item{axes}{Either 'v' or 'h', is the image scaned trough vertically or
horizontally.}

\item{img}{A matrix of imagedata resulting from a readJpeg import.}
}
\description{
\code{isWhite} is looking for the parts in an image that are
white (rgb = 255) and returns a logical vector (T or F).
}
\examples{
whiteVector <- isWhite(letter, axes = 'h')
}

