% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rotateAndCombine}
\alias{rotateAndCombine}
\title{Rotate Letter randomly}
\usage{
rotateAndCombine(letter, angle, nrows = 80, ncols = 800)
}
\arguments{
\item{letter}{}

\item{angle}{}

\item{nrows}{}

\item{ncols}{}
}
\description{
\code{rotateAndCombine}
}

