% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sharpenImg}
\alias{sharpenImg}
\title{Makes dark colors black and light colors white, i.e. sharpens the colors}
\usage{
sharpenImg(img, limit)
}
\arguments{
\item{img}{An image resulting from a readImage import.}

\item{limit}{The color limit value [0, 1], where 0 is black and 1 is white.}
}
\description{
\code{sharpenImg} transforms a multilayered and multicolored
Image into a Greyscale image.
}

