% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotLetters}
\alias{plotLetters}
\title{Plot Letter on a canvas}
\usage{
plotLetters(letter, canvas, number, letter_width = 30)
}
\arguments{
\item{letter}{}

\item{canvas}{}

\item{number}{}

\item{letter_width}{}
}
\description{
\code{plotLetters}
}

